AT2/requirements.txt
aiohttp==3.6.1
Appium-Python-Client==0.47
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
baidu-aip==2.2.17.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==1.11
django-pdfkit==0.3.1
djangorestframework==3.10.2
dukpy==0.2.2
dwebsocket==0.5.11
fake-useragent==0.1.11
ffmpy3==0.2.3
future==0.17.1
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.23
javascripthon==0.10
Jinja2==2.10.1
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
lxml==4.4.1
macropy3==1.1.0b2
MarkupSafe==1.1.1
memory-profiler==0.55.0
more-itertools==7.2.0
multidict==4.5.2
numpy==1.17.2
opencv-python==3.4.2.16
packaging==19.2
pdfkit==0.6.0
Pillow==6.1.0
pluggy==0.13.0
prettytable==0.7.2
psutil==5.6.3
py==1.8.0
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pygame==1.9.6
pyparsing==2.4.2
pytesseract==0.3.0
pytest==5.2.2
python-docx==0.8.10
python-resources==0.3
pytz==2019.2
PyYAML==5.1.2
reportlab==3.5.26
requests==2.22.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
typing-extensions==3.7.4
urllib3==1.25.4
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0
